Split monolithic component into a presentational and a container one.
This is to seperate the concerns of fetching the data and visualizing it.
