+++ client/components/tinymce/plugins/wpcom-view/wpvideo-view.jsx
@@ -1,8 +1,7 @@
 /**
  * External dependencies
  */
-import React, { Component } from 'react';
-import defaults from 'lodash/object/defaults';
+import React, { Component, PropTypes } from 'react';
 import omitBy from 'lodash/omitBy';
 import find from 'lodash/collection/find';
 import QueryString from 'querystring';
@@ -12,76 +11,13 @@ import tinymce from 'tinymce/tinymce';
  * Internal dependencies
  */
 import shortcodeUtils from 'lib/shortcode';
-import MediaActions from 'lib/media/actions';
-import MediaListStore from 'lib/media/list-store';
 import MediaUtils from 'lib/media/utils';
-
-const getStateData = siteId => {
-	return {
-		media: MediaListStore.getAll( siteId ),
-		mediaHasNextPage: MediaListStore.hasNextPage( siteId ),
-		mediaFetchingNextPage: MediaListStore.isFetchingNextPage( siteId )
-	};
-};
 
 class WpVideoView extends Component {
 
-	static match( content ) {
-		const match = shortcodeUtils.next( 'wpvideo', content );
-
-		if ( match ) {
-			return {
-				index: match.index,
-				content: match.content,
-				options: {
-					shortcode: match.shortcode
-				}
-			};
-		}
-	}
-
-	static serialize( content ) {
-		return encodeURIComponent( content );
-	}
-
-	constructor( props ) {
-		super( props );
-		this.state = getStateData( props.siteId );
-	}
-
-	componentDidMount() {
-		MediaListStore.on( 'change', this.updateStateData.bind( this ) );
-		this.fetchData( this.props.siteId );
-	}
-
-	componentWillUnmount() {
-		MediaListStore.off( 'change', this.updateStateData.bind( this ) );
-	}
-
-	componentWillReceiveProps( nextProps ) {
-		if ( this.props.siteId !== nextProps.siteId ) {
-			this.fetchData( this.props.siteId );
-			this.updateStateData();
-		}
-	}
-
-	fetchData( siteId ) {
-		setTimeout( () => {
-			MediaActions.setQuery( siteId, { mime_type: 'video/' } );
-			MediaActions.fetchNextPage( siteId )
-		}, 0 );
-	}
-
-	updateStateData() {
-		this.setState( getStateData( this.props.siteId ) );
-		if ( ! this.state.mediaFetchingNextPage && this.state.mediaHasNextPage ) {
-			this.fetchData( this.props.siteId );
-		}
-	}
-
 	getVideoAttributes( videopress_guid ) {
-		if ( this.state.media ) {
-			return find( this.state.media, item => MediaUtils.isVideoPressItem( item ) && item.videopress_guid === videopress_guid );
+		if ( this.props.media ) {
+			return find( this.props.media, item => MediaUtils.isVideoPressItem( item ) && item.videopress_guid === videopress_guid );
 		}
 	}
 
@@ -175,5 +111,10 @@ class WpVideoView extends Component {
 	}
 
 }
+
+WpVideoView.propTypes = {
+	media: PropTypes.array,
+	content: PropTypes.string
+};
 
 export default WpVideoView;
